finish backend

+++ backend/controllers/goalControllers.js
@@ -1,11 +1,14 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@desc Get goals 
 //@route Get /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({
+        user: req.user.id
+    })
 
     res.status(200).json(goals);
 })
@@ -19,7 +22,8 @@ const setGoals = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field')  
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal)
 })
@@ -33,6 +37,19 @@ const updateGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error ('Goal not found ')
     }
+
+    const user = await User.findById(req.user.id)
+    //check for user 
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make Sure the logged in user macthed the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorized')
+    }
+
     const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new : true,
     })
@@ -47,6 +64,18 @@ const deleteGoals = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error ('Goal not found ')
+    }
+
+    const user = await User.findById(req.user.id)
+    //check for user 
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make Sure the logged in user macthed the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorized')
     }
     const updateGoal = await Goal.findByIdAndRemove(req.params.id, req.body, {
         new : true,
